requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.8.0
loguru==0.7.2
openai==1.55.1
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
